+++ src/block.cpp
@@ -12,7 +12,7 @@
 #include "common.hpp"
 #include "hashtable.hpp"
 
-Block::Block(const std::vector<uint8_t> data, uint16_t width, uint16_t height)
+Block::Block(const std::vector<uint8_t> data, uint32_t width, uint32_t height)
     : m_width(width), m_height(height), m_picked_strategy(HORIZONTAL) {
   for (size_t i = 0; i < N_STRATEGIES; i++) {
     m_data[i].reserve(width * height);
@@ -21,7 +21,7 @@ Block::Block(const std::vector<uint8_t> data, uint16_t width, uint16_t height)
   m_strategy_results.fill({0, 0});
 }
 
-Block::Block(uint16_t width, uint16_t height, SerializationStrategy strategy)
+Block::Block(uint32_t width, uint32_t height, SerializationStrategy strategy)
     : m_width(width), m_height(height), m_picked_strategy(strategy) {
 }
