+++ src/block_writer.cpp
@@ -62,8 +62,8 @@ void flush_bits_to_file(std::ofstream& file) {
 }
 
 // Function to write tokens to binary file with bit packing
-bool write_blocks_to_stream(const std::string& filename, uint16_t width,
-                            uint16_t height, uint16_t offset_length,
+bool write_blocks_to_stream(const std::string& filename, uint32_t width,
+                            uint32_t height, uint16_t offset_length,
                             uint16_t length_bits, bool adaptive, bool model,
                             const std::vector<Block>& blocks) {
   std::ofstream file(filename, std::ios::binary);
